+++ src/main/kotlin/com/github/andreypfau/intellijton/func/language.kt
@@ -10,6 +10,7 @@ object FuncFileType : LanguageFileType(FuncLanguage) {
     override fun getName() = "func"
     override fun getDescription() = "FunC language file"
     override fun getDefaultExtension() = "func"
+    val extensions = setOf("func", "fc")
     override fun getIcon() = FuncIcons.FILE
     override fun getCharset(file: VirtualFile, content: ByteArray?) = Charsets.UTF_8.name()
 }