+++ src/main/kotlin/com/github/andreypfau/intellijton/func/psi/factory.kt
@@ -10,12 +10,16 @@ val Project.funcPsiFactory get() = FuncPsiFactory(this)
 
 class FuncPsiFactory(val project: Project) {
 
+    fun createFile(code: CharSequence): FuncFile = createFromText(code) ?: error("Failed to create file: `$code`")
+
     fun createIdentifier(name: String): PsiElement =
         createFromText<FuncFunction>("() $name();")?.functionName?.identifier
             ?: error("Failed to create identifier: `$name`")
 
-    private inline fun <reified T : FuncElement> createFromText(code: CharSequence): T? =
+    inline fun <reified T : FuncElement> createFromText(code: CharSequence): T? = createFromText(code, T::class.java)
+
+    fun <T : FuncElement> createFromText(code: CharSequence, type: Class<T>): T? =
         PsiFileFactory.getInstance(project)
             .createFileFromText("DUMMY.${FuncFileType.defaultExtension}", FuncFileType, code)
-            .childOfType()
+            .childOfType(type)
 }